feat: better api (#24)

+++ src/main/java/com/manager/payments/adapter/out/persistence/payments/PaymentRepositoryAdapter.java
@@ -1,16 +1,19 @@
 package com.manager.payments.adapter.out.persistence.payments;
 
 import com.manager.payments.application.port.out.PaymentRepository;
+import com.manager.payments.model.exceptions.PaymentNotFoundException;
 import com.manager.payments.model.payments.Payment;
 import com.manager.payments.model.payments.PaymentStatus;
-import org.springframework.stereotype.Component;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-@Component
+@Repository
 public class PaymentRepositoryAdapter implements PaymentRepository {
 
     private final PaymentMapper paymentMapper;
@@ -20,6 +23,12 @@ public class PaymentRepositoryAdapter implements PaymentRepository {
                                     PaymentJpaRepository paymentJpaRepository) {
         this.paymentMapper = paymentMapper;
         this.paymentJpaRepository = paymentJpaRepository;
+    }
+
+    @Override
+    public Page<Payment> findAll(Pageable pageable) {
+        Page<PaymentJpaEntity> paymentJpaEntities = paymentJpaRepository.findAll(pageable);
+        return paymentJpaEntities.map(paymentMapper::toPayment);
     }
 
     @Override
@@ -44,12 +53,20 @@ public class PaymentRepositoryAdapter implements PaymentRepository {
     }
 
     @Override
+    public boolean existsById(UUID id) {
+        return paymentJpaRepository.existsById(id);
+    }
+
+    @Override
     public boolean existsByCode(String code) {
         return paymentJpaRepository.existsByCode(code);
     }
 
     @Override
     public void deleteById(UUID id) {
+        if (!paymentJpaRepository.existsById(id))
+            throw new PaymentNotFoundException(id);
+
         paymentJpaRepository.deleteById(id);
     }
